+++ .prettierrc.js
@@ -1,8 +1,0 @@
-
-module.exports = {
-	semi: true,
-	trailingComma: "all",
-	singleQuote: false,
-	printWidth: 120,
-	tabWidth: 4
-  };